fix(rms_variants_train): overhead-bench fp16 dtype + report regime/dedup
Two defects found when aggregating the full Phase 3 sweep:

1. norm_overhead_bench.py: the gradient touch-term `tf.reduce_sum(grads[0])`
   is fp32 while `loss` is fp16 under mixed_float16 -> AddV2 dtype clash
   crashed every fp16 bench row. Cast the term to loss dtype. The step-2
   unit tests missed this because they ran --no-fp16 only.

2. report.py headline aggregation grouped by (experiment, norm, mode) but
   ignored `regime`, pooling stress-regime cells (lr_high collapse, etc.)
   into the headline average -- and the merged all_runs.csv double-counts
   the default regime when an experiment runs in both the core chunk and
   the regime chunk. Unequal n then broke paired-test pairing and mislabeled
   unpaired rows as FAIL. Fix: restrict headline/verdicts/recommendation to
   regime=default and deduplicate on (experiment,norm,mode,seed). Stress
   regimes remain analysed in regime_delta_summary.csv.

+++ src/train/rms_variants_train/norm_overhead_bench.py
@@ -154,7 +154,9 @@ def _time_norm(
             # Touch the gradients so TF actually computes them in tf.function.
             if grads is None or len(grads) == 0:
                 return loss
-            return loss + tf.reduce_sum(grads[0]) * 0.0
+            # Cast the gradient touch-term to loss dtype: under mixed_float16
+            # `loss` is fp16 while `grads[0]` may be fp32 (AddV2 dtype clash).
+            return loss + tf.cast(tf.reduce_sum(grads[0]), loss.dtype) * 0.0
 
         # Warmup
         for _ in range(max(1, warmup)):
